Downloads/Mini-project-main/Mini-project-main/pyend/requirements.txt
Flask
flask-cors
google
requests
beautifulsoup4
yfinance
pandas
python-dotenv
ollama
